[GN-128] ActivityDetailPage 클라이언트 컴포넌트로 분리 및 동적 로딩 처리
